default to voucher, refactor paywall for store
