+++ wondrous-bot-admin/src/pages/store/index.tsx
@@ -13,9 +13,9 @@ import { PricingOptionsTitle, getPlan } from "components/Pricing/PricingOptionsL
 import useAlerts, { usePaywall, useSubscription } from "utils/hooks";
 import PageSpinner from "components/PageSpinner";
 import StoreModule from "components/ModulesActivation/StoreModule";
+import { useStorePaywall } from "utils/storeUtils";
 
 const StorePage = () => {
-  const [isActivateModuleModalOpen, setIsActivateModuleModalOpen] = useState(false);
   const { activeOrg } = useContext(GlobalContext);
   const [openPublishModal, setOpenPublishModal] = useState(false);
   const navigate = useNavigate();
@@ -25,12 +25,7 @@ const StorePage = () => {
     notifyOnNetworkStatusChange: true,
   });
 
-  const {setSnackbarAlertMessage, setSnackbarAlertOpen, setSnackbarAlertAnchorOrigin} = useAlerts();
-  const subscription = useSubscription();
-
-  const { setPaywall, setPaywallMessage, setOnCancel, setCanBeClosed } = usePaywall();
-
-  const plan = getPlan(subscription?.tier);
+  const { isActivateModuleModalOpen, handleSuccess } = useStorePaywall();
 
   useEffect(() => {
     if (activeOrg?.id) {
@@ -46,47 +41,15 @@ const StorePage = () => {
     }
   }, [activeOrg?.id]);
 
-  useEffect(() => {
-    if (plan === PricingOptionsTitle.Basic) {
-      setPaywall(true);
-      setCanBeClosed(false);
-      setPaywallMessage("You discovered a paid feature! Upgrade to unlock it.");
-      setOnCancel(() => {
-        return () => {
-          setPaywall(false);
-          setPaywallMessage("");
-          navigate("/");
-        };
-      });
-    }
-    if (
-      (plan === PricingOptionsTitle.Premium || plan === PricingOptionsTitle.Ecosystem) &&
-      !activeOrg?.modules?.cmtyStore
-    ) {
-      setIsActivateModuleModalOpen(true);
-    }
-  }, [plan, activeOrg]);
-
   const handleNavigationToNewProduct = () => {
     navigate("/store/items/create");
   };
 
-  const handleSuccess = () => {
-    setSnackbarAlertMessage("Community store enabled!")
-    setSnackbarAlertAnchorOrigin({
-      vertical: 'bottom',
-      horizontal: 'right'
-    })
-    setSnackbarAlertOpen(true)
-    setIsActivateModuleModalOpen(false)
-  };
   if (!activeOrg) return null;
 
   return (
     <>
-      {isActivateModuleModalOpen ? (
-        <StoreModule onSuccess={handleSuccess} onCancel={() => navigate("/")} />
-      ) : null}
+      {isActivateModuleModalOpen ? <StoreModule onSuccess={handleSuccess} onCancel={() => navigate("/")} /> : null}
       <PublishDirectoryModal openPublishModal={openPublishModal} setOpenPublishModal={setOpenPublishModal} />
       <PageHeader
         title={`${data?.getStoreItemsForOrg?.length || 0} Products`}
